refactor: remove unused utils
